+++ editor.h
@@ -7,16 +7,16 @@
 
 typedef struct {
     Buffer model;
-    size_t scroll_row;
-    size_t scroll_col;
+    int scroll_row;
+    int scroll_col;
     int show_line_numbers;
     int syntax_highlight;
-    size_t cursor_line;
-    size_t cursor_col;
-    size_t selection_start_line;
-    size_t selection_start_col;
-    size_t selection_end_line;
-    size_t selection_end_col;
+    int cursor_line;
+    int cursor_col;
+    int selection_start_line;
+    int selection_start_col;
+    int selection_end_line;
+    int selection_end_col;
     int selection_active;
     char search_buffer[256];
     int search_mode;
